Add some async test code, and prove we can do printing
Note: Actual printing using the print APIs is a shit-show; it's slow and crashy and some quick googling suggests it is not robust.  However, we can use a canvas (offscreen, in this example, but it could be on-screen) to very quickly generate a PNG that we can open in the system viewer for printing.  That seems like a reasonable approach.